Condition the usage of default key and input key
We've introduced the idea of the key to Unspent/TxIn, and now we look
for `not None` and thusly branch of this.  Previously we introduced the
default_ variables which are the passed in private_key (thus why it has
no default_ prefix).

We isolated the setting of variables to the top of the iteration loop,
and only make variable name changes below.

One possible improvement would be to eliminate the default key,
scriptCode, and public key (and each of their lens) at the top, and
delay that until iteration.  I shied away from this due to comments
about speed.